pyproject.toml
[tool.poetry]
name = "d4-bot-vacations"
version = "0.1.0"
description = ""
authors = ["Vladislav Akashev <vlad.akashev.w@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mailru-im-bot = "^0.0.21"
python-dotenv = "^1.0.1"
transitions = "^0.9.1"
pymongo = "^4.8.0"
pydantic = "^2.8.2"
redis = "^5.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
